--- core/src/main/java/io/grpc/InternalBinaryLogs.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2018, gRPC Authors All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.grpc;
-
-@Internal
-public final class InternalBinaryLogs {
-  public static <ReqT, RespT> ServerMethodDefinition<?, ?> wrapMethodDefinition(
-      BinaryLog binaryLog, ServerMethodDefinition<ReqT, RespT> oMethodDef) {
-    return binaryLog.surrogate.wrapMethodDefinition(oMethodDef);
-  }
-
-  public static Channel wrapChannel(BinaryLog binaryLog, Channel channel) {
-    return binaryLog.surrogate.wrapChannel(channel);
-  }
-
-  public static BinaryLog createBinaryLog(BinaryLogProvider surrogate) {
-    return new BinaryLog(surrogate);
-  }
-
-  private InternalBinaryLogs() {}
-}
