+++ services/src/main/java/io/grpc/services/BinaryLogs.java
@@ -18,17 +18,16 @@ package io.grpc.services;
 
 import io.grpc.BinaryLog;
 import io.grpc.ExperimentalApi;
-import io.grpc.InternalBinaryLogs;
 import java.io.IOException;
 
 @ExperimentalApi("https://github.com/grpc/grpc-java/issues/4017")
 public final class BinaryLogs {
   public static BinaryLog createBinaryLog() throws IOException {
-    return InternalBinaryLogs.createBinaryLog(new BinaryLogProviderImpl());
+    return new BinaryLogProviderImpl();
   }
 
   public static BinaryLog createCensusBinaryLog() throws IOException {
-    return InternalBinaryLogs.createBinaryLog(new CensusBinaryLogProvider());
+    return new CensusBinaryLogProvider();
   }
 
   private BinaryLogs() {}
